+++ CHANGELOG.md
@@ -8,11 +8,11 @@
   * add locale support
   * add more functions to binary to get a proper CLI tool (or move to adyen-admin??)
   * more helper e.g. cutom fields
-  * read adyen-admin credentials from cfg file??
   * auto re-login when session timed out
-  * show local / remote skins, option to sync
 
 ## v0.3.0
+  * read adyen-admin credentials from cfg file
+  * show local / remote skins, option to sync
   * add skin upload -> 'adyen-admin'
   * add skin export
   * gemify